Remove unused variables

+++ mpi/abstract_comm.hpp
@@ -451,7 +451,7 @@ public:
 		finish_count += finish_count__;
 
 		while(finish_count > 0) {
-			if(empty_list.size() == MAX_REQUESTS) {
+                  if(empty_list.size() == (long unsigned int)MAX_REQUESTS) {
 				fprintf(IMD_OUT, "Error: No active request\n");
 				throw "Error: No active request";
 			}
